moved jw_get_gaussian_states to utils/sparse_tools.py
