+++ src/openfermion/utils/_sparse_tools.py
@@ -28,7 +28,8 @@ from openfermion.config import *
 from openfermion.ops import (FermionOperator, hermitian_conjugated,
                              normal_ordered, number_operator,
                              QuadraticHamiltonian, QubitOperator)
-from openfermion.utils import commutator, fourier_transform, Grid
+from openfermion.utils import (commutator, fourier_transform,
+                               gaussian_state_preparation_circuit, Grid)
 from openfermion.hamiltonians._jellium import (momentum_vector,
                                                position_vector,
                                                grid_indices)
@@ -367,6 +368,104 @@ def jw_get_ground_states_by_particle_number(sparse_operator, particle_number,
     return ground_energy, ground_states
 
 
+def jw_get_gaussian_state(quadratic_hamiltonian, occupied_orbitals=None):
+    """Compute an eigenvalue and eigenstate of a quadratic Hamiltonian.
+
+    Eigenstates of a quadratic Hamiltonian are also known as fermionic
+    Gaussian states.
+
+    Args:
+        quadratic_hamiltonian(QuadraticHamiltonian):
+            The Hamiltonian whose eigenstate is desired.
+        occupied_orbitals(list):
+            A list of integers representing the indices of the occupied
+            orbitals in the desired Gaussian state. If this is None
+            (the default), then it is assumed that the ground state is
+            desired, i.e., the orbitals with negative energies are filled.
+
+    Returns
+    -------
+        energy (float):
+            The eigenvalue.
+        state (sparse):
+            The eigenstate in scipy.sparse csc format.
+    """
+    if not isinstance(quadratic_hamiltonian, QuadraticHamiltonian):
+        raise ValueError('Input must be an instance of QuadraticHamiltonian.')
+
+    n_qubits = quadratic_hamiltonian.n_qubits
+
+    # Compute the energy
+    orbital_energies, constant = quadratic_hamiltonian.orbital_energies()
+    if occupied_orbitals is None:
+        # The ground energy is desired
+        if quadratic_hamiltonian.conserves_particle_number:
+            num_negative_energies = numpy.count_nonzero(
+                    orbital_energies < -EQ_TOLERANCE)
+            occupied_orbitals = range(num_negative_energies)
+        else:
+            occupied_orbitals = []
+    energy = numpy.sum(orbital_energies[occupied_orbitals]) + constant
+
+    # Obtain the circuit that prepares the Gaussian state
+    circuit_description, start_orbitals = gaussian_state_preparation_circuit(
+            quadratic_hamiltonian, occupied_orbitals)
+
+    # Initialize the starting state
+    state = jw_slater_determinant(start_orbitals, n_qubits)
+
+    # Apply the circuit
+    particle_hole_transformation = (
+            jw_sparse_particle_hole_transformation_last_mode(n_qubits))
+    for parallel_ops in circuit_description:
+        for op in parallel_ops:
+            if op == 'pht':
+                state = particle_hole_transformation.dot(state)
+            else:
+                i, j, theta, phi = op
+                state = jw_sparse_givens_rotation(
+                            i, j, theta, phi, n_qubits).dot(state)
+
+    return energy, state
+
+
+def jw_sparse_givens_rotation(i, j, theta, phi, n_qubits):
+    """Return the matrix (acting on a full wavefunction) that performs a
+    Givens rotation of modes i and j in the Jordan-Wigner encoding."""
+    if j != i + 1:
+        raise ValueError('Only adjacent modes can be rotated.')
+    if j > n_qubits - 1:
+        raise ValueError('Too few qubits requested.')
+
+    cosine = numpy.cos(theta)
+    sine = numpy.sin(theta)
+    phase = numpy.exp(1.j * phi)
+
+    # Create the two-qubit rotation matrix
+    rotation_matrix = scipy.sparse.csc_matrix(
+            ([1., phase * cosine, -phase * sine, sine, cosine, phase],
+             ((0, 1, 1, 2, 2, 3), (0, 1, 2, 1, 2, 3))),
+            shape=(4, 4))
+
+    # Initialize identity operators
+    left_eye = scipy.sparse.eye(2 ** i, format='csc')
+    right_eye = scipy.sparse.eye(2 ** (n_qubits - 1 - j), format='csc')
+
+    # Construct the matrix and return
+    givens_matrix = kronecker_operators([left_eye, rotation_matrix, right_eye])
+
+    return givens_matrix
+
+
+def jw_sparse_particle_hole_transformation_last_mode(n_qubits):
+    """Return the matrix (acting on a full wavefunction) that performs a
+    particle-hole transformation on the last mode in the Jordan-Wigner
+    encoding.
+    """
+    left_eye = scipy.sparse.eye(2 ** (n_qubits - 1), format='csc')
+    return kronecker_operators([left_eye, pauli_matrix_map['X']])
+
+
 def get_density_matrix(states, probabilities):
     n_qubits = states[0].shape[0]
     density_matrix = scipy.sparse.csc_matrix(
@@ -386,34 +485,24 @@ def is_hermitian(sparse_operator):
     return True
 
 
-def get_ground_state(operator):
+def get_ground_state(sparse_operator):
     """Compute lowest eigenvalue and eigenstate.
-
     Returns:
         eigenvalue: The lowest eigenvalue, a float.
         eigenstate: The lowest eigenstate in scipy.sparse csc format.
     """
-    if isinstance(operator, QuadraticHamiltonian):
-        from openfermion.utils._slater_determinants import (
-                jw_get_gaussian_state)
-        eigenvalue, eigenstate = jw_get_gaussian_state(operator)
-    elif isinstance(operator, scipy.sparse.spmatrix):
-        if not is_hermitian(operator):
-            raise ValueError('operator must be Hermitian.')
+    if not is_hermitian(sparse_operator):
+        raise ValueError('sparse_operator must be Hermitian.')
 
-        values, vectors = scipy.sparse.linalg.eigsh(
-            operator, 2, which='SA', maxiter=1e7)
+    values, vectors = scipy.sparse.linalg.eigsh(
+        sparse_operator, 2, which='SA', maxiter=1e7)
 
-        order = numpy.argsort(values)
-        values = values[order]
-        vectors = vectors[:, order]
-        eigenvalue = values[0]
-        eigenstate = scipy.sparse.csc_matrix(vectors[:, 0]).T
-    else:
-        raise ValueError('operator must be a sparse matrix or '
-                         'QuadraticHamiltonian.')
-
-    return eigenvalue, eigenstate
+    order = numpy.argsort(values)
+    values = values[order]
+    vectors = vectors[:, order]
+    eigenvalue = values[0]
+    eigenstate = scipy.sparse.csc_matrix(vectors[:, 0])
+    return eigenvalue, eigenstate.T
 
 
 def sparse_eigenspectrum(sparse_operator):
